+++ src/message/extension.rs
@@ -1,6 +1,9 @@
 use crate::buffer::Buf;
+use arrayvec::ArrayVec;
 use nom::{bytes::complete::take, number::complete::be_u16, IResult};
 use std::ops::Range;
+
+pub type ExtensionVec = ArrayVec<Extension, { ExtensionType::supported().len() }>;
 
 #[derive(Debug, PartialEq, Eq, Default)]
 pub struct Extension {
@@ -191,13 +194,13 @@ impl ExtensionType {
         Ok((input, ExtensionType::from_u16(value)))
     }
 
-    /// Returns true if this extension type is a known/supported variant.
-    pub fn is_known(&self) -> bool {
-        Self::all().contains(self)
+    /// Returns true if this extension type is supported by this implementation.
+    pub fn is_supported(&self) -> bool {
+        Self::supported().contains(self)
     }
 
-    /// All known extension types that this implementation handles.
-    pub fn all() -> &'static [ExtensionType] {
+    /// Supported extension types that this implementation handles.
+    pub const fn supported() -> &'static [ExtensionType; 8] {
         &[
             ExtensionType::SupportedGroups,
             ExtensionType::EcPointFormats,